requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
python-dotenv==1.0.1
Werkzeug==3.0.1
google-auth==2.27.0
PyJWT==2.8.0
requests==2.31.0
flask==3.0.0
python-dotenv==1.0.0
groq==0.4.1
jsonschema==4.20.0
requests==2.31.0
flask
flask-sqlalchemy
flask-jwt-extended
flask-cors
python-dotenv

